% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\docType{data}
\name{disney_colours}
\alias{disney_colours}
\title{The list of disney colours}
\format{An object of class \code{character} of length 45.}
\usage{
disney_colours
}
\description{
The list of disney colours
}
\keyword{datasets}
