% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\name{disney_cols}
\alias{disney_cols}
\title{Function to extract colours as hex codes}
\usage{
disney_cols(...)
}
\arguments{
\item{...}{Character names of disney_colours}
}
\description{
Function to extract colours as hex codes
}
