% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\name{disney_pal}
\alias{disney_pal}
\title{Return function to interpolate a disney colour palette}
\usage{
disney_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in disney_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a disney colour palette
}
