% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\docType{data}
\name{disney_palettes}
\alias{disney_palettes}
\title{Function to create palettes}
\format{An object of class \code{list} of length 7.}
\usage{
disney_palettes
}
\description{
Function to create palettes
}
\keyword{datasets}
