% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\docType{data}
\name{nature_colours}
\alias{nature_colours}
\title{The list of nature colours}
\format{An object of class \code{character} of length 110.}
\usage{
nature_colours
}
\description{
The list of nature colours
}
\keyword{datasets}
