% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\name{nature_cols}
\alias{nature_cols}
\title{Function to extract colours as hex codes}
\usage{
nature_cols(...)
}
\arguments{
\item{...}{Character names of nature_colours}
}
\description{
Function to extract colours as hex codes
}
