% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\name{nature_pal}
\alias{nature_pal}
\title{Return function to interpolate a nature colour palette}
\usage{
nature_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nature_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a nature colour palette
}
