% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\docType{data}
\name{nature_palettes}
\alias{nature_palettes}
\title{Function to create palettes}
\format{An object of class \code{list} of length 13.}
\usage{
nature_palettes
}
\description{
Function to create palettes
}
\keyword{datasets}
