% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\name{scale_colour_disney}
\alias{scale_colour_disney}
\title{Colour scale constructor for some disney-ish colours}
\usage{
scale_colour_disney(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in disney_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
Colour scale of disney-like palette
}
\description{
This is the default colour scale for categorical variables for the disney-like palette.
It does not generate colour-blind safe palettes.
These are the palettes to choose from:
\itemize{
   \item \code{cinderella}
   \item \code{monet}
   \item \code{small_world}
   \item \code{alice}
   \item \code{pan}
   \item \code{when_i_was_your_age}
   \item \code{firefly}
   \item \code{main}
} \cr
Palette Colours inspired by: \cr
    http://elijahmeeks.com/#content/blog/2015_08_17_palettes \cr \cr
Palette made ffg blog post: \cr
    https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2 \cr
}
\examples{
# Colour using the small_world palette
library(ggplot2)
ggplot(mtcars, aes(hp, mpg, colour = cyl)) +
  geom_point(size = 4, alpha = .8) +
  scale_colour_disney(
    discrete = FALSE,
    palette = "small_world",
    guide = "none"
  ) +
  facet_wrap(~cyl)

# Colour using the `alice` palette, and where cyl is discrete
library(ggplot2)
ggplot(mtcars, aes(hp, mpg, colour = as.factor(cyl))) +
  geom_point(size = 4, alpha = .8) +
  scale_colour_disney(
    discrete = TRUE,
    palette = "alice",
    guide = "none"
  ) +
  facet_wrap(~cyl)
}
