% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenityr.R
\name{scale_colour_nature}
\alias{scale_colour_nature}
\title{Colour scale constructor for some nature colours}
\usage{
scale_colour_nature(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nature_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
Colour scale of natural palette
}
\description{
This is the default colour scale for categorical variables for the nature palette.
It does not generate colour-blind safe palettes.
These are the palettes to choose from:
\itemize{
   \item \code{uyuni}
   \item \code{okavango}
   \item \code{lakelouise}
   \item \code{provence}
   \item \code{halong}
   \item \code{vatnajokull}
   \item \code{arashiyama}
   \item \code{mountcook}
   \item \code{benagil}
   \item \code{bryce}
   \item \code{jozi}
   \item \code{firefly}
   \item \code{main}
} \cr
Palette Colours inspired by: \cr
    https://medium.com/sketch-app-sources/10-color-palettes-from-the-natural-world-to-inspire-your-creative-streak-bc2fb73d161e  \cr \cr
Palette made ffg blog post: \cr
    https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2 \cr
}
\examples{
# Colour using the `okavango` palette
ggplot(mtcars, aes(hp, mpg, colour = cyl)) +
  geom_point(size = 4, alpha = .8) +
  scale_colour_nature(
    discrete = FALSE,
    palette = "uyuni",
    guide = "none"
  ) +
  facet_wrap(~cyl) +
  theme_dark()

# Colour using the `halong` palette, and where cyl is discrete
ggplot(mtcars, aes(hp, mpg, colour = as.factor(cyl))) +
  geom_point(size = 4, alpha = .8) +
  scale_colour_nature(
    discrete = TRUE,
    palette = "halong",
    guide = "none"
  ) +
  facet_wrap(~cyl)
}
