% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\name{scale_fill_nature}
\alias{scale_fill_nature}
\title{Fill scale constructor for some nature colours}
\usage{
scale_fill_nature(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in nature_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
Fill scale of nature palette
}
\description{
This is the default fill scale for the nature palette.
It does not generate colour-blind safe palettes.
These are the palettes to choose from:
\itemize{
   \item \code{uyuni}
   \item \code{okavango}
   \item \code{lakelouise}
   \item \code{provence}
   \item \code{halong}
   \item \code{vatnajokull}
   \item \code{arashiyama}
   \item \code{mountcook}
   \item \code{benagil}
   \item \code{bryce}
   \item \code{jozi}
   \item \code{main}
} \cr
Palette Colors inspired by: \cr
    https://medium.com/sketch-app-sources/10-color-palettes-from-the-natural-world-to-inspire-your-creative-streak-bc2fb73d161e  \cr \cr
Palette made ffg blog post: \cr
    https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2 \cr
}
\examples{
# Fill by discrete variable with different palette + remove legend (guide)
library(ggplot2)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_nature(palette = "vatnajokull", guide = "none")
}
