% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werpals.R
\name{werpals_display}
\alias{werpals_display}
\title{Creates RColorBrewer like palette viewers for the disney and nature palettes}
\usage{
werpals_display(pal)
}
\arguments{
\item{pal}{Character name of palette to display colours for - disney or nature, here}
}
\value{
A heatmap of the colours for the disney and nature palettes
}
\description{
Creates RColorBrewer like palette viewers for the disney and nature palettes
}
\examples{
# View the disney colours
werpals_display(pal = "disney")
# View the nature colours
werpals_display(pal = "nature")
}
